refactor(form-field): convert to typescript
