+++ src/__internal__/form-field/form-field.style.ts
@@ -1,0 +1,29 @@
+import styled, { css } from "styled-components";
+import { space } from "styled-system";
+import { baseTheme } from "../../style/themes";
+
+const FormFieldStyle = styled.div`
+  & + & {
+    margin-top: 16px;
+  }
+
+  &&& {
+    ${space}
+  }
+`;
+
+FormFieldStyle.defaultProps = {
+  theme: baseTheme,
+};
+
+export interface FieldLineStyleProps {
+  inline?: boolean;
+}
+const FieldLineStyle = styled.div<FieldLineStyleProps>`
+  ${({ inline }) => css`
+    display: ${inline ? "flex" : "block"};
+  `}
+`;
+
+export { FieldLineStyle };
+export default FormFieldStyle;
