--- src/__internal__/form-field/index.d.ts
@@ -1,5 +1,0 @@
-export {
-  default,
-  CommonFormFieldPropTypes,
-  FormFieldPropTypes,
-} from "./form-field";
