--- src/__internal__/form-field/index.js
@@ -1,1 +1,0 @@
-export { default } from "./form-field.component";
