+++ src/__internal__/form-field/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./form-field.component";
+export type { FormFieldProps } from "./form-field.component";
